org.apache.servicecomb.samples.practise.houserush.customer.manage.aggregate.User move org.apache.servicecomb.samples.practise.houserush.customer.manage.rpc.po.User
