+++ houserush/customer-manage/src/main/java/org/apache/servicecomb/samples/practise/houserush/customer/manage/rpc/po/User.java
@@ -15,44 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.samples.practise.houserush.customer.manage.aggregate;
+package org.apache.servicecomb.samples.practise.houserush.customer.manage.rpc.po;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import javax.persistence.*;
-import java.util.Date;
 
 @Data
 public class User {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.AUTO)
-  private Integer id;
-
   private String username;
 
-  @Transient
   private String password;
-
-  @JsonIgnore
-  private String hashedPassword;
-
-  @Temporal(TemporalType.TIMESTAMP)
-  private Date deletedAt;
-
-  @CreatedDate
-  @Temporal(TemporalType.TIMESTAMP)
-  private Date createdAt;
-
-  @LastModifiedDate
-  @Temporal(TemporalType.TIMESTAMP)
-  private Date updatedAt;
-
-  @Transient
-  private String token;
-
 }
